Move exposed ownership pieces to swift/ABI/
We've been exposing all of `swift/AST/Ownership.h` to functions in
`swift/ABI` since 2018. It looks like that was done only for the
definition of the `ValueOwnership` enum. This commit pulls that enum out
into the ABI directory to make it clearer that this is exposed to the
Swift ABI (via MetadataValues.h) and removed the AST header from
MetadataValues.h to prevent accidents.