Basic: add a workaround for VS2019
When building with MSVC, we seem to be hitting a crash in the compiler.
The crash seems to originate from the use of the `member_constructible`
predicate being used in the `std::enable_if`.  Use a local variable
instead to avoid this issue.  This problem is already resolved in
VS2022, and hopefully we will be switching to that soon.