prevent uses of `Copyable`; fix bogus errors when `~Copyable` appears
Rather than try to omit explicit uses of `Copyable` in the output module
for backwards compatability, we instead make it an error to use `Copyable`
as a type when not suppressed. While it's straightforward to omit it for
generic parameters, for uses of `Copyable` as an existential, we'd need to
emit `Any` in its place. We'd also need to find and omit or modify
explicit type constraints or same-type requirements involving Copyable.

All of that sounded rather gross, and not really needed right now since
any uses of `Copyable` explicitly is just not needed.

update test with Copyable being banned generally