[SemanticARCOpts] Load_borrow moved load [copy]s.
In the context of promoting a `load [copy]` to a `load_borrow`, allow
the value representing the lifetime of the load to differ from the `load
[copy]` itself--it may either be the load or its only user.

If promoting the `load [copy]` to a `load_borrow` with the load itself
as the lifetime representative fails, look for a single `move_value` use
of the load.  If found, attempt the optimization again with the
`move_value` as the lifetime representative.

rdar://108514447