[CanonBorrowScope] Allow inner val without outer.
When rewriting outer uses of the borrow, `recursivelyRewriterOuterUses`
is called recursively.  First, it is called with the borrow itself.
Subsequently, it is called with forwarding users of the borrow.  Such a
user may not be rewriteable as outer uses.  If it isn't
`innerToOuterMap` won't have an entry for it.

Previously, failing to find an entry would raise an assertion.  Here,
the function is made to return a "null" SILValue.  Its caller
RewriteOuterBorrowUses::visitForwardingUse was already written to handle
such a result.