Remove an obsolete workaround for error variables.
This is a revert  of the workaround for creating debug  info for error variables
in b2109ab4db10, while leaving the test  added back then in place.  The compiler
is now emitting debug info for the error pattern binding as it's supposed to and
after the  recent migration to  stricter debug  scope generation, there  are now
situations where the  variable added for the workaround and  the correct one are
in conflict.

rdar://108576484
(cherry picked from commit 06682897449d2e8b4b52bfdfcb35b380c319d055)