Convert the noimplicitcopy addressonly fail test back to its original form.
For some reason in a previous commit, this started to work so I just added
support. In this PR, it changed again and the test started to fail again. We are
not shipping noimplicitcopy in 5.9, so I just changed the error message to the
old "I don't understand" error message.

I additionally looked at why the failure is occuring. The reason why is that the
frontend emits an unchecked_ref_cast since currently the
moveonlywrapper_to_copyable instruction does not accept addresses (even though
it could) and the field sensitive liveness analysis does not know how to
represent an unchecked_ref_cast (which is ok, since we would like it to be
rejected).

(cherry picked from commit d25cd0bce215472bee7c569187fb12f65be7484c)