[Sema] Erase existential PATs to reduced type
When opening an existential return or parameter type that has a
primaray associated type, the primmary asssociated type should be
preserved if the type can be reduced to a concrete type. Otherwise,
we should erase the existential to a type with no dependent members
or generic parameters and diagnose explicit coercion is needed.