[test] Take care of change out of deleted directories
These two tests were changing directories into the temporary, never
changing out of there, and then recreating the temporary, which left the
rest of the test executing in a non-existing directory.

Modify those tests to change out to the previous directory and avoid
executing any command in a non-existing directory.