[stdlib] StringProtocol._ephemeralString: Make inlinable & force-inline
At this point, this is just an alias of the generic String.init.
Don’t make it slower than that.

Foundation is currently using this to implement its StringProtocol
APIs. If we ever introduce a shared string form, making this inlinable
will not interfere with Foundation reaping the benefits of that,
because Foundation is always built with the latest stdlib.

(But it seems better to have Foundation normalize `StringProtocol`
values to `Substring` rather than `String`. This entry point seems
like the wrong abstraction to me.)

Also, stop using `_ephemeralString` within the stdlib.